pyproject.toml
[project]
name = "kh-instability"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Jonathan Elsner", email = "jeelsner@outlook.com"},
]
dependencies = ["pillow>=11.0.0", "numpy>=2.1.3", "scipy>=1.14.1", "matplotlib>=3.9.2", "ffmpeg-python>=0.2.0"]
requires-python = ">=3.12"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true

[dependency-groups]
dev = [
    "pytest>=8.3.3",
    "coverage>=7.6.7",
    "ruff>=0.8.0",
]
interactive = [
    "jupyter>=1.1.1",
    "ipympl>=0.9.4",
]
